minor: refactor tests
