+++ tests/Unit/History/SnapshotTest.php
@@ -9,13 +9,12 @@
  * file that was distributed with this source code.
  */
 
-namespace Zenstruck\Messenger\Monitor\Tests\History;
+namespace Zenstruck\Messenger\Monitor\Tests\Unit\History;
 
 use PHPUnit\Framework\TestCase;
 use Zenstruck\Messenger\Monitor\History\Snapshot;
 use Zenstruck\Messenger\Monitor\History\Specification;
 use Zenstruck\Messenger\Monitor\History\Storage;
-
 use function Zenstruck\collect;
 
 /**
