+++ tests/Unit/TransportsTest.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Zenstruck\Messenger\Monitor\Tests;
+namespace Zenstruck\Messenger\Monitor\Tests\Unit;
 
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Cache\Adapter\NullAdapter;
@@ -17,9 +17,9 @@ use Symfony\Component\DependencyInjection\ServiceLocator;
 use Symfony\Component\Messenger\MessageBusInterface;
 use Symfony\Component\Messenger\Transport\Sync\SyncTransport;
 use Symfony\Component\Messenger\Transport\TransportInterface;
-use Zenstruck\Messenger\Monitor\Tests\Fixture\Stub\CountableListableTransport;
-use Zenstruck\Messenger\Monitor\Tests\Fixture\Stub\CountableTransport;
-use Zenstruck\Messenger\Monitor\Tests\Fixture\Stub\ListableTransport;
+use Zenstruck\Messenger\Monitor\Tests\Unit\Stub\CountableListableTransport;
+use Zenstruck\Messenger\Monitor\Tests\Unit\Stub\CountableTransport;
+use Zenstruck\Messenger\Monitor\Tests\Unit\Stub\ListableTransport;
 use Zenstruck\Messenger\Monitor\Transports;
 use Zenstruck\Messenger\Monitor\Worker\WorkerCache;
 use Zenstruck\Messenger\Monitor\Workers;
