+++ tests/bootstrap.php
@@ -1,0 +1,7 @@
+<?php
+
+use Symfony\Component\Filesystem\Filesystem;
+
+require \dirname(__DIR__).'/vendor/autoload.php';
+
+(new Filesystem())->remove(__DIR__.'/../var');
